Address comments and fix travis failure
