+++ gobblin-runtime/src/test/java/org/apache/gobblin/runtime/spec_catalog/FlowCatalogTest.java
@@ -15,16 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.gobblin.spec_catalog;
+package org.apache.gobblin.runtime.spec_catalog;
 
+import com.google.common.base.Optional;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.typesafe.config.Config;
 import java.io.File;
 import java.net.URI;
 import java.util.Collection;
 import java.util.Properties;
-
 import org.apache.commons.io.FileUtils;
 import org.apache.gobblin.runtime.api.FlowSpec;
+import org.apache.gobblin.runtime.api.Spec;
 import org.apache.gobblin.runtime.api.SpecExecutor;
+import org.apache.gobblin.runtime.api.SpecNotFoundException;
+import org.apache.gobblin.runtime.app.ServiceBasedAppLauncher;
+import org.apache.gobblin.runtime.spec_executorInstance.InMemorySpecExecutor;
+import org.apache.gobblin.util.ConfigUtils;
+import org.apache.gobblin.util.PathUtils;
 import org.apache.hadoop.fs.Path;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -33,33 +42,20 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
-import com.google.common.base.Optional;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.typesafe.config.Config;
 
-import org.apache.gobblin.runtime.api.Spec;
-import org.apache.gobblin.runtime.api.TopologySpec;
-import org.apache.gobblin.runtime.app.ServiceBasedAppLauncher;
-import org.apache.gobblin.runtime.spec_catalog.TopologyCatalog;
-import org.apache.gobblin.runtime.spec_executorInstance.InMemorySpecExecutor;
-import org.apache.gobblin.util.ConfigUtils;
-import org.apache.gobblin.util.PathUtils;
-
-
-public class TopologyCatalogTest {
-
-  private static final Logger logger = LoggerFactory.getLogger(TopologyCatalog.class);
+public class FlowCatalogTest {
+  private static final Logger logger = LoggerFactory.getLogger(FlowCatalog.class);
   private static Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
   private static final String SPEC_STORE_PARENT_DIR = "/tmp";
-  private static final String SPEC_STORE_DIR = "/tmp/topologyTestSpecStore";
-  private static final String SPEC_DESCRIPTION = "Test Topology Spec";
+  private static final String SPEC_STORE_DIR = "/tmp/flowTestSpecStore";
+  private static final String SPEC_GROUP_DIR = "/tmp/flowTestSpecStore/flowTestGroupDir";
+  private static final String SPEC_DESCRIPTION = "Test Flow Spec";
   private static final String SPEC_VERSION = FlowSpec.Builder.DEFAULT_VERSION;
 
   private ServiceBasedAppLauncher serviceLauncher;
-  private TopologyCatalog topologyCatalog;
-  private TopologySpec topologySpec;
+  private FlowCatalog flowCatalog;
+  private FlowSpec flowSpec;
 
   @BeforeClass
   public void setup() throws Exception {
@@ -71,33 +67,44 @@ public class TopologyCatalogTest {
     Properties properties = new Properties();
     properties.put("specStore.fs.dir", SPEC_STORE_DIR);
 
-    this.serviceLauncher = new ServiceBasedAppLauncher(properties, "TopologyCatalogTest");
+    this.serviceLauncher = new ServiceBasedAppLauncher(properties, "FlowCatalogTest");
 
-    this.topologyCatalog = new TopologyCatalog(ConfigUtils.propertiesToConfig(properties),
+    this.flowCatalog = new FlowCatalog(ConfigUtils.propertiesToConfig(properties),
         Optional.of(logger));
-    this.serviceLauncher.addService(topologyCatalog);
+    this.serviceLauncher.addService(flowCatalog);
 
     // Start Catalog
     this.serviceLauncher.start();
 
     // Create Spec to play with
-    this.topologySpec = initTopologySpec();
+    this.flowSpec = initFlowSpec(SPEC_STORE_DIR);
   }
 
-  private TopologySpec initTopologySpec() {
+  /**
+   * Create FlowSpec with default URI
+   */
+  public static FlowSpec initFlowSpec(String specStore) {
+    return initFlowSpec(specStore, computeFlowSpecURI());
+  }
+
+  /**
+   * Create FLowSpec with specified URI and SpecStore location.
+   */
+  public static FlowSpec initFlowSpec(String specStore, URI uri){
     Properties properties = new Properties();
-    properties.put("specStore.fs.dir", SPEC_STORE_DIR);
+    properties.put("specStore.fs.dir", specStore);
     properties.put("specExecInstance.capabilities", "source:destination");
     Config config = ConfigUtils.propertiesToConfig(properties);
 
     SpecExecutor specExecutorInstanceProducer = new InMemorySpecExecutor(config);
 
-    TopologySpec.Builder topologySpecBuilder = TopologySpec.builder(computeTopologySpecURI())
+    FlowSpec.Builder flowSpecBuilder = null;
+    flowSpecBuilder = FlowSpec.builder(uri)
         .withConfig(config)
         .withDescription(SPEC_DESCRIPTION)
         .withVersion(SPEC_VERSION)
-        .withSpecExecutor(specExecutorInstanceProducer);
-    return topologySpecBuilder.build();
+        .withTemplate(URI.create("templateURI"));
+    return flowSpecBuilder.build();
   }
 
   @AfterClass
@@ -112,59 +119,64 @@ public class TopologyCatalogTest {
   }
 
   @Test
-  public void createTopologySpec() {
+  public void createFlowSpec() {
     // List Current Specs
-    Collection<Spec> specs = topologyCatalog.getSpecs();
+    Collection<Spec> specs = flowCatalog.getSpecs();
     logger.info("[Before Create] Number of specs: " + specs.size());
     int i=0;
     for (Spec spec : specs) {
-      TopologySpec topologySpec = (TopologySpec) spec;
-      logger.info("[Before Create] Spec " + i++ + ": " + gson.toJson(topologySpec));
+      FlowSpec flowSpec = (FlowSpec) spec;
+      logger.info("[Before Create] Spec " + i++ + ": " + gson.toJson(flowSpec));
     }
     Assert.assertTrue(specs.size() == 0, "Spec store should be empty before addition");
 
     // Create and add Spec
-    this.topologyCatalog.put(topologySpec);
+    this.flowCatalog.put(flowSpec);
 
     // List Specs after adding
-    specs = topologyCatalog.getSpecs();
+    specs = flowCatalog.getSpecs();
     logger.info("[After Create] Number of specs: " + specs.size());
     i = 0;
     for (Spec spec : specs) {
-      topologySpec = (TopologySpec) spec;
-      logger.info("[After Create] Spec " + i++ + ": " + gson.toJson(topologySpec));
+      flowSpec = (FlowSpec) spec;
+      logger.info("[After Create] Spec " + i++ + ": " + gson.toJson(flowSpec));
     }
     Assert.assertTrue(specs.size() == 1, "Spec store should contain 1 Spec after addition");
   }
 
-  @Test (dependsOnMethods = "createTopologySpec")
-  public void deleteTopologySpec() {
+  @Test (dependsOnMethods = "createFlowSpec")
+  void testExist() throws Exception {
+    Assert.assertTrue(flowCatalog.exists(flowSpec.getUri()));
+  }
+
+  @Test (dependsOnMethods = "testExist")
+  public void deleteFlowSpec() throws SpecNotFoundException {
     // List Current Specs
-    Collection<Spec> specs = topologyCatalog.getSpecs();
+    Collection<Spec> specs = flowCatalog.getSpecs();
     logger.info("[Before Delete] Number of specs: " + specs.size());
     int i=0;
     for (Spec spec : specs) {
-      TopologySpec topologySpec = (TopologySpec) spec;
-      logger.info("[Before Delete] Spec " + i++ + ": " + gson.toJson(topologySpec));
+      FlowSpec flowSpec = (FlowSpec) spec;
+      logger.info("[Before Delete] Spec " + i++ + ": " + gson.toJson(flowSpec));
     }
     Assert.assertTrue(specs.size() == 1, "Spec store should initially have 1 Spec before deletion");
 
-    this.topologyCatalog.remove(topologySpec.getUri());
+    this.flowCatalog.remove(flowSpec.getUri());
 
     // List Specs after adding
-    specs = topologyCatalog.getSpecs();
-    logger.info("[After Create] Number of specs: " + specs.size());
+    specs = flowCatalog.getSpecs();
+    logger.info("[After Delete] Number of specs: " + specs.size());
     i = 0;
     for (Spec spec : specs) {
-      topologySpec = (TopologySpec) spec;
-      logger.info("[After Create] Spec " + i++ + ": " + gson.toJson(topologySpec));
+      flowSpec = (FlowSpec) spec;
+      logger.info("[After Delete] Spec " + i++ + ": " + gson.toJson(flowSpec));
     }
     Assert.assertTrue(specs.size() == 0, "Spec store should be empty after deletion");
   }
 
-  public URI computeTopologySpecURI() {
+  public static URI computeFlowSpecURI() {
     // Make sure this is relative
-    URI uri = PathUtils.relativizePath(new Path(SPEC_STORE_DIR), new Path(SPEC_STORE_PARENT_DIR)).toUri();
+    URI uri = PathUtils.relativizePath(new Path(SPEC_GROUP_DIR), new Path(SPEC_STORE_PARENT_DIR)).toUri();
     return uri;
   }
 }
