+++ gobblin-runtime/src/test/java/org/apache/gobblin/runtime/spec_catalog/TopologyCatalogTest.java
@@ -15,27 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.gobblin.spec_catalog;
+package org.apache.gobblin.runtime.spec_catalog;
 
-import com.google.common.base.Optional;
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-import com.typesafe.config.Config;
 import java.io.File;
 import java.net.URI;
-import java.net.URISyntaxException;
 import java.util.Collection;
 import java.util.Properties;
+
 import org.apache.commons.io.FileUtils;
 import org.apache.gobblin.runtime.api.FlowSpec;
-import org.apache.gobblin.runtime.api.Spec;
 import org.apache.gobblin.runtime.api.SpecExecutor;
-import org.apache.gobblin.runtime.api.SpecNotFoundException;
-import org.apache.gobblin.runtime.app.ServiceBasedAppLauncher;
-import org.apache.gobblin.runtime.spec_catalog.FlowCatalog;
-import org.apache.gobblin.runtime.spec_executorInstance.InMemorySpecExecutor;
-import org.apache.gobblin.util.ConfigUtils;
-import org.apache.gobblin.util.PathUtils;
 import org.apache.hadoop.fs.Path;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -44,20 +33,33 @@ import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
 
+import com.google.common.base.Optional;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.typesafe.config.Config;
 
-public class FlowCatalogTest {
-  private static final Logger logger = LoggerFactory.getLogger(FlowCatalog.class);
+import org.apache.gobblin.runtime.api.Spec;
+import org.apache.gobblin.runtime.api.TopologySpec;
+import org.apache.gobblin.runtime.app.ServiceBasedAppLauncher;
+import org.apache.gobblin.runtime.spec_catalog.TopologyCatalog;
+import org.apache.gobblin.runtime.spec_executorInstance.InMemorySpecExecutor;
+import org.apache.gobblin.util.ConfigUtils;
+import org.apache.gobblin.util.PathUtils;
+
+
+public class TopologyCatalogTest {
+
+  private static final Logger logger = LoggerFactory.getLogger(TopologyCatalog.class);
   private static Gson gson = new GsonBuilder().setPrettyPrinting().create();
 
   private static final String SPEC_STORE_PARENT_DIR = "/tmp";
-  private static final String SPEC_STORE_DIR = "/tmp/flowTestSpecStore";
-  private static final String SPEC_GROUP_DIR = "/tmp/flowTestSpecStore/flowTestGroupDir";
-  private static final String SPEC_DESCRIPTION = "Test Flow Spec";
+  private static final String SPEC_STORE_DIR = "/tmp/topologyTestSpecStore";
+  private static final String SPEC_DESCRIPTION = "Test Topology Spec";
   private static final String SPEC_VERSION = FlowSpec.Builder.DEFAULT_VERSION;
 
   private ServiceBasedAppLauncher serviceLauncher;
-  private FlowCatalog flowCatalog;
-  private FlowSpec flowSpec;
+  private TopologyCatalog topologyCatalog;
+  private TopologySpec topologySpec;
 
   @BeforeClass
   public void setup() throws Exception {
@@ -69,44 +71,33 @@ public class FlowCatalogTest {
     Properties properties = new Properties();
     properties.put("specStore.fs.dir", SPEC_STORE_DIR);
 
-    this.serviceLauncher = new ServiceBasedAppLauncher(properties, "FlowCatalogTest");
+    this.serviceLauncher = new ServiceBasedAppLauncher(properties, "TopologyCatalogTest");
 
-    this.flowCatalog = new FlowCatalog(ConfigUtils.propertiesToConfig(properties),
+    this.topologyCatalog = new TopologyCatalog(ConfigUtils.propertiesToConfig(properties),
         Optional.of(logger));
-    this.serviceLauncher.addService(flowCatalog);
+    this.serviceLauncher.addService(topologyCatalog);
 
     // Start Catalog
     this.serviceLauncher.start();
 
     // Create Spec to play with
-    this.flowSpec = initFlowSpec(SPEC_STORE_DIR);
+    this.topologySpec = initTopologySpec();
   }
 
-  /**
-   * Create FlowSpec with default URI
-   */
-  public static FlowSpec initFlowSpec(String specStore) {
-    return initFlowSpec(specStore, computeFlowSpecURI());
-  }
-
-  /**
-   * Create FLowSpec with specified URI and SpecStore location.
-   */
-  public static FlowSpec initFlowSpec(String specStore, URI uri){
+  private TopologySpec initTopologySpec() {
     Properties properties = new Properties();
-    properties.put("specStore.fs.dir", specStore);
+    properties.put("specStore.fs.dir", SPEC_STORE_DIR);
     properties.put("specExecInstance.capabilities", "source:destination");
     Config config = ConfigUtils.propertiesToConfig(properties);
 
     SpecExecutor specExecutorInstanceProducer = new InMemorySpecExecutor(config);
 
-    FlowSpec.Builder flowSpecBuilder = null;
-    flowSpecBuilder = FlowSpec.builder(uri)
+    TopologySpec.Builder topologySpecBuilder = TopologySpec.builder(computeTopologySpecURI())
         .withConfig(config)
         .withDescription(SPEC_DESCRIPTION)
         .withVersion(SPEC_VERSION)
-        .withTemplate(URI.create("templateURI"));
-    return flowSpecBuilder.build();
+        .withSpecExecutor(specExecutorInstanceProducer);
+    return topologySpecBuilder.build();
   }
 
   @AfterClass
@@ -121,64 +112,59 @@ public class FlowCatalogTest {
   }
 
   @Test
-  public void createFlowSpec() {
+  public void createTopologySpec() {
     // List Current Specs
-    Collection<Spec> specs = flowCatalog.getSpecs();
+    Collection<Spec> specs = topologyCatalog.getSpecs();
     logger.info("[Before Create] Number of specs: " + specs.size());
     int i=0;
     for (Spec spec : specs) {
-      FlowSpec flowSpec = (FlowSpec) spec;
-      logger.info("[Before Create] Spec " + i++ + ": " + gson.toJson(flowSpec));
+      TopologySpec topologySpec = (TopologySpec) spec;
+      logger.info("[Before Create] Spec " + i++ + ": " + gson.toJson(topologySpec));
     }
     Assert.assertTrue(specs.size() == 0, "Spec store should be empty before addition");
 
     // Create and add Spec
-    this.flowCatalog.put(flowSpec);
+    this.topologyCatalog.put(topologySpec);
 
     // List Specs after adding
-    specs = flowCatalog.getSpecs();
+    specs = topologyCatalog.getSpecs();
     logger.info("[After Create] Number of specs: " + specs.size());
     i = 0;
     for (Spec spec : specs) {
-      flowSpec = (FlowSpec) spec;
-      logger.info("[After Create] Spec " + i++ + ": " + gson.toJson(flowSpec));
+      topologySpec = (TopologySpec) spec;
+      logger.info("[After Create] Spec " + i++ + ": " + gson.toJson(topologySpec));
     }
     Assert.assertTrue(specs.size() == 1, "Spec store should contain 1 Spec after addition");
   }
 
-  @Test (dependsOnMethods = "createFlowSpec")
-  void testExist() throws Exception {
-    Assert.assertTrue(flowCatalog.exists(flowSpec.getUri()));
-  }
-
-  @Test (dependsOnMethods = "testExist")
-  public void deleteFlowSpec() throws SpecNotFoundException {
+  @Test (dependsOnMethods = "createTopologySpec")
+  public void deleteTopologySpec() {
     // List Current Specs
-    Collection<Spec> specs = flowCatalog.getSpecs();
+    Collection<Spec> specs = topologyCatalog.getSpecs();
     logger.info("[Before Delete] Number of specs: " + specs.size());
     int i=0;
     for (Spec spec : specs) {
-      FlowSpec flowSpec = (FlowSpec) spec;
-      logger.info("[Before Delete] Spec " + i++ + ": " + gson.toJson(flowSpec));
+      TopologySpec topologySpec = (TopologySpec) spec;
+      logger.info("[Before Delete] Spec " + i++ + ": " + gson.toJson(topologySpec));
     }
     Assert.assertTrue(specs.size() == 1, "Spec store should initially have 1 Spec before deletion");
 
-    this.flowCatalog.remove(flowSpec.getUri());
+    this.topologyCatalog.remove(topologySpec.getUri());
 
     // List Specs after adding
-    specs = flowCatalog.getSpecs();
-    logger.info("[After Delete] Number of specs: " + specs.size());
+    specs = topologyCatalog.getSpecs();
+    logger.info("[After Create] Number of specs: " + specs.size());
     i = 0;
     for (Spec spec : specs) {
-      flowSpec = (FlowSpec) spec;
-      logger.info("[After Delete] Spec " + i++ + ": " + gson.toJson(flowSpec));
+      topologySpec = (TopologySpec) spec;
+      logger.info("[After Create] Spec " + i++ + ": " + gson.toJson(topologySpec));
     }
     Assert.assertTrue(specs.size() == 0, "Spec store should be empty after deletion");
   }
 
-  public static URI computeFlowSpecURI() {
+  public URI computeTopologySpecURI() {
     // Make sure this is relative
-    URI uri = PathUtils.relativizePath(new Path(SPEC_GROUP_DIR), new Path(SPEC_STORE_PARENT_DIR)).toUri();
+    URI uri = PathUtils.relativizePath(new Path(SPEC_STORE_DIR), new Path(SPEC_STORE_PARENT_DIR)).toUri();
     return uri;
   }
 }
